move and export `CallbackError`
