+++ packages/http-server/src/main.ts
@@ -6,6 +6,7 @@
  * @packageDocumentation
  */
 
+export * from './callback-error.js'
 export * from '@tbdex/http-client'
 export * from './http-server.js'
 export * from './types.js'
